requirements.txt
siray>=0.2.1
requests~=2.32.5
pillow~=11.3.0

numpy~=2.0.2
torchvision~=0.23.0